fix : restructure the folders
