+++ src/components/providers/components/next-auth.provider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { SessionProvider} from "next-auth/react"
+import { SessionProvider } from "next-auth/react";
 
-export default function NextAuthProvider ({ children }: { children: React.ReactNode }) {
-    return <SessionProvider> {children} </SessionProvider>
+export default function NextAuthProvider({ children }: { children: React.ReactNode }) {
+    return <SessionProvider> {children} </SessionProvider>;
 }
