--- src/components/providers/components/theme.provider.tsx
@@ -1,11 +1,0 @@
-"use client";
-import * as React from "react";
-import { ThemeProvider as NextThemesProvider } from "next-themes";
-
-export function ThemeProvider({ children, ...props }: React.ComponentProps<typeof NextThemesProvider>) {
-    return (
-        <NextThemesProvider {...props} defaultTheme="dark">
-            {children}
-        </NextThemesProvider>
-    );
-}
