--- src/components/providers/nextIntl-client/components/nextIntl-client.provider.tsx
@@ -1,5 +1,0 @@
-import { NextIntlClientProvider } from "next-intl";
-
-export default function NextIntlProvider({ children }: { children: React.ReactNode }) {
-    return <NextIntlClientProvider>{children}</NextIntlClientProvider>;
-}
