--- src/components/providers/session-provider/index.tsx
@@ -1,13 +1,0 @@
-"use client";
-
-import { SessionProvider } from "next-auth/react";
-
-const ParentSessionProvider = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <>
-      <SessionProvider>{children}</SessionProvider>
-    </>
-  );
-};
-
-export default ParentSessionProvider;
