+++ src/lib/types/auth/auth.d.ts
@@ -5,3 +5,22 @@ export type Step = (typeof FORGOT_PASSWORD_STEPS)[keyof typeof FORGOT_PASSWORD_S
 export type VerifyOtpResponse =
     | { status: "Success" } // success case
     | { error: string }; // error case
+
+declare global {
+    type LoginResponse = {
+        token: string;
+        user: {
+            _id: string;
+            firstName: string;
+            lastName: string;
+            email: string;
+            gender: string;
+            phone: string;
+            photo: string;
+            role: string;
+            wishlist: any[];
+            addresses: string;
+            createdAt: string;
+        };
+    };
+}
