+++ src/lib/types/end-point-api/delete-all-notifications.d.ts
