+++ src/lib/types/end-point-api/delete-notifications.d.ts
