--- src/lib/types/index.d.ts
@@ -1,2 +1,0 @@
-export * from "./end-point-api/products";
-// export * from "./components";
